Change rule for not applicable values
Changing rule for not applicable values: if the term 'opc' is not on the field name, then instead replace the missing value to a negative value ('N') instead of a not applicable value. If the term does appear on the field name, then leave it with the not applicable value.
Also adding transformation to reduced files so that the tactile sensibility values hyperesthesia, anesthesia and hypoesthesia or hyperalgesia, analgesia and hypoalgesia were transformed to "altered".